refactor: move install code to router.ts
